refactor: commanify gstr_utils and make gstr_2 independent
