+++ india_compliance/gst_india/doctype/gstr_import_log/gstr_import_log.py
@@ -73,7 +73,7 @@ def toggle_scheduled_jobs(stopped):
     scheduled_job = frappe.db.get_value(
         "Scheduled Job Type",
         {
-            "method": "india_compliance.gst_india.utils.gstr.download_queued_request",
+            "method": "india_compliance.gst_india.utils.gstr_utils.download_queued_request",
         },
     )
