+++ india_compliance/gst_india/utils/gstr_2/gstr.py
@@ -1,13 +1,9 @@
 import frappe
-from frappe import _
-from frappe.utils import add_to_date, now_datetime
 
 from india_compliance.gst_india.constants import STATE_NUMBERS
 from india_compliance.gst_india.doctype.gst_inward_supply.gst_inward_supply import (
     create_inward_supply,
 )
-from india_compliance.gst_india.utils import get_gstin_list
-from india_compliance.gst_india.utils.gstr import ReturnsAPI
 
 
 def get_mapped_value(value, mapping):
@@ -126,85 +122,3 @@ class GSTR:
 
     def update_gstins(self):
         pass
-
-
-@frappe.whitelist()
-def validate_company_gstins(company=None, company_gstin=None):
-    """
-    Checks the validity of the company's GSTIN authentication.
-
-    Args:
-        company_gstin (str): The GSTIN of the company to validate.
-
-    Returns:
-        dict: A dictionary where the keys are the GSTINs and the values are booleans indicating whether the authentication is valid.
-    """
-    frappe.has_permission("GST Settings", throw=True)
-
-    credentials = get_company_gstin_credentials(company, company_gstin)
-
-    if company_gstin and not credentials:
-        frappe.throw(
-            _("Missing GSTIN credentials for GSTIN: {gstin}.").format(
-                gstin=company_gstin
-            )
-        )
-
-    if not credentials:
-        frappe.throw(_("Missing credentials in GST Settings"))
-
-    if company and not company_gstin:
-        missing_credentials = set(get_gstin_list(company)) - set(
-            credential.gstin for credential in credentials
-        )
-
-        if missing_credentials:
-            frappe.throw(
-                _("Missing GSTIN credentials for GSTIN(s): {gstins}.").format(
-                    gstins=", ".join(missing_credentials),
-                )
-            )
-
-    gstin_authentication_status = {
-        credential.gstin: (
-            credential.session_expiry
-            and credential.auth_token
-            and credential.session_expiry > add_to_date(now_datetime(), minutes=30)
-        )
-        for credential in credentials
-    }
-
-    return gstin_authentication_status
-
-
-def get_company_gstin_credentials(company=None, company_gstin=None):
-    filters = {"service": "Returns"}
-
-    if company:
-        filters["company"] = company
-
-    if company_gstin:
-        filters["gstin"] = company_gstin
-
-    return frappe.get_all(
-        "GST Credential",
-        filters=filters,
-        fields=["gstin", "session_expiry", "auth_token"],
-    )
-
-
-@frappe.whitelist()
-def request_otp(company_gstin):
-    frappe.has_permission("GST Settings", throw=True)
-
-    return ReturnsAPI(company_gstin).request_otp()
-
-
-@frappe.whitelist()
-def authenticate_otp(company_gstin, otp):
-    frappe.has_permission("GST Settings", throw=True)
-
-    api = ReturnsAPI(company_gstin)
-    response = api.autheticate_with_otp(otp)
-
-    return api.process_response(response)
