+++ india_compliance/gst_india/utils/gstr_2/gstr_2a.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import frappe
 
 from india_compliance.gst_india.utils import get_datetime, parse_datetime
-from india_compliance.gst_india.utils.gstr.gstr import GSTR, get_mapped_value
+from india_compliance.gst_india.utils.gstr_2.gstr import GSTR, get_mapped_value
 
 
 def map_date_format(date_str, source_format, target_format):
