+++ india_compliance/gst_india/utils/gstr_2/test_gstr_2b.py
@@ -5,8 +5,8 @@ from frappe import parse_json, read_file
 from frappe.tests.utils import FrappeTestCase
 
 from india_compliance.gst_india.utils import get_data_file_path
-from india_compliance.gst_india.utils.gstr import GSTRCategory, save_gstr_2b
-from india_compliance.gst_india.utils.gstr.test_gstr_2a import TestGSTRMixin
+from india_compliance.gst_india.utils.gstr_2 import GSTRCategory, save_gstr_2b
+from india_compliance.gst_india.utils.gstr_2.test_gstr_2a import TestGSTRMixin
 
 
 class TestGSTR2b(FrappeTestCase, TestGSTRMixin):
