+++ india_compliance/hooks.py
@@ -375,7 +375,7 @@ scheduler_events = {
     "cron": {
         "*/5 * * * *": [
             "india_compliance.gst_india.utils.e_invoice.retry_e_invoice_e_waybill_generation",
-            "india_compliance.gst_india.utils.gstr.download_queued_request",
+            "india_compliance.gst_india.utils.gstr_utils.download_queued_request",
             "india_compliance.gst_india.doctype.purchase_reconciliation_tool.purchase_reconciliation_tool.auto_refresh_authtoken",
         ],
         "0 2 * * *": [
